Introduce a wrapper for arithmetic operations When two numbers are input and are requested to be added or subtracted, the actual arithmetic operation that is to be done on those two operands depends on their individual sign and magnitude.
Example: If a positive number and a negative number are requested to be added
together, then the actual arithmetic operation to be done would be subtraction
and the sign of the result depends on the magnitude of those numbers.

1. For the above reason, I've introduced an operation resolution module which
wraps the underlying functions that do arithmetic and also resolves the exact
arithmetic function call (add or subtract) that is to be called based on the
sign and magnitude of the input operands and the operator.
2. Updated makefile to accommodate the operations module.
3. Updated driver function to test the operations module 'resolve' function.